[Code] Compile projects with -WError
WError / WX was turned of for viewer because it kept showing errors from
the Qt project. This should be fixed in a future update.

A custom enum was used instead of the QMetaType and the QVariantType and
it is easier to use without having to converty between the enums.
Moreover, since the type needed is always known beforehand there was no
need for type deduction from QVariant.